pyproject.toml
[project]
name = "minecraft-purple-agent"
version = "0.1.0"
description = "LLM-powered Minecraft purple agent for AgentBeats MCU benchmark"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.5",
    "anthropic>=0.49.0",
    "openai>=1.60.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
